Extract interface from LineageTreeLayout
There are already three implementations of the LineageTreeLayout:
* LineageTreeLayoutImp
* LongEdgeLineageTreeLayout
* ContextLayout

It is reasonable to introduce an interface. We could especially use the
factory pattern for create LineageTreeLayout.
